+++ src/Crowdin.Api/AI/AiPromptCompletion.cs
@@ -15,7 +15,7 @@ namespace Crowdin.Api.AI
         public OperationStatus Status { get; set; }
         
         [JsonProperty("progress")]
-        public int Progress { get; set; }
+        public long Progress { get; set; }
 
         [JsonProperty("attributes")]
         public AttributesObject Attributes { get; set; }
@@ -36,7 +36,7 @@ namespace Crowdin.Api.AI
         public class AttributesObject
         {
             [JsonProperty("aiPromptId")]
-            public int AiPromptId { get; set; }
+            public long AiPromptId { get; set; }
         }
     }
 }
