+++ src/Crowdin.Api/AI/AiPromptContextResources.cs
@@ -17,41 +17,41 @@ namespace Crowdin.Api.AI
     public class PreTranslateActionAiPromptContextResources : AiPromptContextResources
     {
         [JsonProperty("projectId")]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
 
         [JsonProperty("targetLanguageId")]
         public string TargetLanguageId { get; set; } = null!;
 
         [JsonProperty("stringIds")]
-        public ICollection<int> StringIds { get; set; } = null!;
+        public ICollection<long> StringIds { get; set; } = null!;
     }
 
     [PublicAPI]
     public class AssistActionAiPromptContextResources : AiPromptContextResources
     {
         [JsonProperty("projectId")]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
 
         [JsonProperty("targetLanguageId")]
         public string TargetLanguageId { get; set; } = null!;
 
         [JsonProperty("stringIds")]
-        public ICollection<int> StringIds { get; set; } = null!;
+        public ICollection<long> StringIds { get; set; } = null!;
         
         [JsonProperty("filteredStringsIds")]
-        public ICollection<int>? FilteredStringsIds { get; set; }
+        public ICollection<long>? FilteredStringsIds { get; set; }
     }
 
     [PublicAPI]
     public class CustomActionAiPromptContextResources : AiPromptContextResources
     {
         [JsonProperty("projectId")]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
 
         [JsonProperty("targetLanguageId")]
         public string TargetLanguageId { get; set; } = null!;
 
         [JsonProperty("stringIds")]
-        public ICollection<int> StringIds { get; set; } = null!;
+        public ICollection<long> StringIds { get; set; } = null!;
     }
 }
